mod: Update terraform-plugin-sdk to version v2
Updates the hashicorp terraform SDK module to the new `v2` verison which
introduces a few changes, most of them should be addressed in follow-up
PRs since it would make this PR significantly bigger.

Namely, the module version bump to `v2` which needs to be intrpolated in
all the import paths, and the changes the way the `*api.API` instance is
obtained.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>