terraform: Add support for 0.13 version
Adds support for terraform 0.13, which introduces a few changes on how
third party providers need to be declared in the `.tf` files. Namely, it
modifies the `make install` target so that the binary is placed both on
the new directory structure and the old one at the top level.

Additionally, it modifies the examples and the example in the README to
include the `elastic/ec` source.

If the `required_providers` block is not specified, terraform v0.13 will
attempt to download the terraform binary from:
`registry.terraform.io/hashicorp/ec`.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>