ec_deployment_traffic_filter: Add new resource (#52)
Adds a new resource `ec_deployment_traffic_filter` and modifies the
existing `ec_deployment` resource to add a new field `traffic_filter`,
where traffic filter IDs can be specified.

The patch has been split in three different commits

#### Add the `ec_deployment_traffic_filter` resource (35a28d9)

Adds the resource and makes the field `include_by_default` default to
false like we would in our `cloud-sdk-go` when not specified, rather
than a required field, like the API does.

#### Add `traffic_filter` to `ec_deployment` (b760a31)

The new `traffic_filter` is an optional field of type Set with the hash
algorithm "HashString", it's a list of unique strings which reference a
number of traffic_filtering rules.

On the implementation, there's two ways the traffic_filter rules have to
be handled. 
On Creation, theres a `settings` field on the `CreateRequest`, which is
not present on the `UpdateRequest`, Create specifies a list of traffic 
filters by ID which are added to the deployment on creation.
On Update, there's no `settings` field so we need to rely on the traffic
filter associations which the API present.
On Delete, there's also the challenge of deleting the existing traffic
filter associations before the deployment is Shut down.

To be able to handle the Update and Delete, `handleTrafficFilterChange`
is introduced which handles, the addition and the removal of items from
that `traffic_filter` list.

`handleTrafficFilterChange` checks if there's a `traffic_filter` change
and adds and removes associations of a deployment to a traffic filter
rule set.

Direct use of the V1 APIs without a wrapper package is necessary here
since there's no actual implementation of these APIs on the cloud-sdk-go
(yet). A follow-up PR will address this point.

Last, a `hasDeploymentChange` function has been created to only update
the deployment through PUT with the current state when there's a change
in any field != `traffic_filter` this avoids creating unnecessary no-op
plans on the backend which slow down the changes and clutter the plan
history.

#### Add acceptance tests for both resources (4bca40e)

Adds the integration tests which handle the CRUD of the new resource
`ec_deployment_traffic_filter`, the case where a deployment is assigned
new `traffic_filter` rules and `ec_deployment_traffic_filter` sweepers.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>