ec_stack: Add Data Source
Adds a new data source which allows the terraform provider users to get
information about a specific stack version by using a `version_regex`.
The `version_regex` field also accepts a `"latest"` string as a value
which will cause the data source to obtain the latest available stack
version and return it, best used with `lock`.

There's quite a lot of fields contained in the Stack models and this PR
implements most of the fields which might not even be useful in the long
run, but can be useful to browse around what capabilities are in a stack
version.

```hcl
data "ec_stack" "latest" {
  version_regex = "latest"
  region        = "us-east-1"
  lock          = true
}
```

```hcl
data "ec_stack" "latest" {
  version_regex = "7.9.?"
  region        = "us-east-1"
}
```

Signed-off-by: Marc Lopez <marc5.12@outlook.com>