ec_deployments: New datasource (#88)
Adds a new data source which allows the terraform provider users to
retrieve a list of deployments based on a specified query defined by
pre-set filters.

Many more filters can be added in the future, but for now I'd like to
keep it simple by having only the ones that are most likely to be used
(based on feedback from internal users).

I have added acceptance testing as part of the same test with the
ec_deployment datasource. Given it's such a small test, it made
sense to me to have it this way for test speed, and to avoid spinning up
more instances than needed.