ec_resource: Use deployment template as blueprint (#98)
Changes the `topology` block within each of deployment resources from
`required` to `optional` by using the `deployment_template` template of
the `GET deployment templates` API call. Not only aligning with how the
Elastic Cloud Web UI works, but also benefiting from defaults, important
metadata and pre-configured settings set in the deployment template now
consumed, as a result a series of _side-effects_ have cascaded down to
the `ec_resource` schema itself, allowing users to create deployments
specifying much less, and making getting started much easier.

Both on `Create` and `Update`, the `GET /deployments/templates` endpoint
is called, filtering out the specified `deployment_template_id` field,
and reading the `.deployment_template` object which contains a blueprint
to create a deployment using this template.
The `deployment_template` contains all the resource kinds node_types,
curation settings, node attributes (for tiered data topologies), and
`instance_configuration_id`, all details that the user was burdened to
know before to be able to create a deployment. Since the Elasticsearch
resource is the most complex of resources, it's the one benefiting most
from this change.
Then, the template is used as a blueprint, and any attributes set by the
user will be overridden from the deployment template blueprint.

Also removing the default values for `memory_per_node` and `zone_count`.
This caused a little bit more of an impact on the code than expected,
thus all the handling of empty and / or unset topology settings and mem
size.

Since there's no reason to allow users to tweak the Elasticsearch and
Enterprise Search `node_type_*`, they have been removed as settable, and 
are only computed for reference.

As part of this patch, the "region" field has been marked as `ForceNew`
since changing the region field will cause a deployment to be destroyed
and a new one created.

Additionally, two different bugs are fixed in this patch.

First, removing any resource from the Terraform definition didn't cause
the resource to be removed from the backend due to `prune_orphans` being
set to `false`.

Second, any resources which were removed, still appear on the plan with
"stopped" in the `status`, since no filtering was done on the flattening
functions, an infinite loop was triggered and they were attempted to be
removed again, and persisted in the state. Filtering any resources with
`status == "stopped"` fixes this.

Last, a series of new unit tests have been added which now test more
scenarios where users are some overrides on the topology to ensure that
the code behaves as it should.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>