ec_resource: Fix overwritten elasticsearch config
This patch fixes a bug where the topology.ElasticsearchConfig setting
was being overridden when some topology settings were being set, since
there was a reassignment back to `elem.Elasticsearch`.

The change now takes in the existing ElasticsearchConfig and sets what
needs to be overwritten instead of creating a new object and reassigning
the result back to the property, which was creating this issue.

It is unlikely that the same situation occurs in the other deployment
resources, but it might be good to change the other resources functions
use the same signature.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>