api client: Enable HTTP retries on timeout
Enables performing HTTP request retries upon receiving a timeout from
the API. This should increase the reliability of any of the HTTP calls
performed by the provider.

Sets the amount of retries to `2`, meaning that any request which times
out, will be retries for up to 2 times using the default back-off time.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>