acceptance: Make test fields dynamic (#121)
All region, deployment_template_id and instance_configuration_id fields
in the acceptance tests are now dynamic. Region fields are set through a
new environment variable EC_REGION, or set to a default if this variable
is not set.

Only ESS regions are supported. ECE environments are not supported as
the deployment template names and configurations will vary depending
on the ECE installation.

Additionally, the files within the testdata/ directory have been modified
to use default values rather specifying topology when it was not necessary
to test this way. This reduced complexity in our test templates.