ec_deployment: Support legacy plans through import
Adds all of the necessary pieces too allow "legacy" (pre-ECE v2 and very
old ESS deployments which haven't changed in a looong time) deployments
to be managed by terraform through `terraform import`.

The first problem is that the `"region"` and `"version"` weren't being
set to the plan on the read, which caused the whole deployment to be
tainted when imported (region changes from `""` => `"<value>"`, causing
the whole deployment to be re-created (ForceNew field in effect).

Second, `deploymentapi.Get` needed `ConvertLegacyPlans` set to `true`.

Last, the deployment topology elements were being ignored since the size
is `0` because `memory_per_node` and `node_count_per_zone` are used in
the legacy plan definitions. These changes only needed to be made in the
`elasticsearch` and `kibana` resource kinds since they're the only ones
that existed pre-DNT.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>