ec_deployment: Fix broken user_settings_json field
Fixes `user_settings_json` and `user_settings_overrides_json` content
parsing in the expanders and flatteners where the value of the field was
being set as is instead of being handled correctly through the `json`
library which Marshals / Unmarshals the contents of the field.

Additionally, a little bit of a cleanup to provide consistent behavior
for the config expanders has been done too, where the destination struct
is passed as a pointer instead of being returned, and now an error is
returned (should any `json` encoding / decoding occur).

Furthermore, the `region` and `version` field have been added to the
flatteners causing the import acceptance tests to not have to ignore the
fields anymore.

Last, a few test fixtures and cases for different templates have been
added as well, including a case in the acceptance tests where the apm
resource has a `user_settings_json` set, since json and yaml settings
cannot be included at the same time.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>