acceptance: Remove acceptance build tag
Removes the acceptance build tag and instead adds a new function which
calls `t.Skip()` when a function with a `*testing.T` as the first arg
requires external HTTP access, in which case it is skipped, to trigger
this behavior any tests which fall into that category should call
`requiresAPIConn` at the beginning of the function to ensure that we're
not incurring any errors when calling `newAPI()`.

This was bothering me since writing tests was becoming quite cumbersome
due to the go build avoidance on the `+build acceptance` tag.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>