acceptance: Fix ec_deployment tests (#159)
As there was an active incident yesterday which was affecting deployment
creation, the TestAccDeployment_basic_defaults_hw test broke and
went unnoticed. This PR fixes said test.

Additionally, I am adding a separate .tf configuration file for that test
so any changes meant for a different test will not affect it.