release: Add goreleaser config and github action
Adds the goreleaser config and release github action for pushes of `v*`
tags. Adds three new targets: `tag`, `snapshot` and `release`.
The `release` target has been added as a fallback to perform a manual
release in case the GitHub action fails.

Last, it sets the version to `v0.1.0-beta` as agreed with the team.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>