release: Add Jenkins release job and file
Adds a new Jenkins job which should be triggered every time a `v*`  tag
is pushed to the repository.

The job will import the GPG private key, export the required environment
variables and run the `goreleaser` binary to trigger a new release. The
GPG key passphrase will be cached within GPG so `goreleaser` can sign
the release without interactively asking for the passphrase.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>