testacc: Sweep all existing 1h+ old deployments
Changes the deployment sweep logic to be time based, deleting all the
deployments that are 1h or older since their last change.
The change will remove any deployments found in the account with the
`terraform_acc_` prefix, without checking the status of any of the
deployment resources.

This results in a proper cleanup of the dev account since the account
had a lot of dangling resources which were costing costing us a few $.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>