ec_traffic_filter: Fix rule ordering bug
Fixes a bug where multiple having a traffic filter with a few rules will
cause an infinite diff loop by changing the field from a List to a Set.
This consistently hashes the list using the same algorithm, which does
not happen when the field is of `schema.TypeList`.

Adds unit and acceptance tests covering this case.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>