ec_deployments: Bugfix for queries with hyphens (#241)
Fixes a bug where the name prefix query was using a text field rather
than a keyword field in order to ensure the query is not analyzed.

Additionally the acceptance tests have been fixed since recently it
appears the deployment read API returns all possible topology elements
even if the size is set to 0. This means that the ec_deployment datasource
will return more topology elements than what is set in the ec_deployment
resource. I will open up a follow up PR with the fix for this, but I wanted to
leave the tests passing for now.