ec_deployment: Fix where datasource was adding unused topology elements (#242)
This PR is a follow up to the bug discovered in #241

The deployment read API returns all possible topology elements
even if the size is set to 0. This means that the ec_deployment datasource
will return more topology elements than what is set in the ec_deployment
resource. This patch makes sure only the topology elements used in the 
deployment are returned.