ec_deployment: add support for data tiers (#253)
Adds support for `node_roles` and automatic migration from `node_type_*`
to `node_roles`. I will note a few things to make it easier to break
down all of these changes.

Adds `node_roles` support by updating the SDK to the latest available
version, this makes the deployment template not work out of the box and
instead our client has to choose whether to use the `node_type` object
or the `node_roles` array.
Stack versions >= `7.10.0` can and should use `node_roles` since that
allows users to leverage the `cold` tier and autoscaling in the future
as well.
With the introduction of the `"id"` field, the topology elements needn't
use `instance_configuration_id` as the unique field, instead, `id` is
that unique field value.
**The `instance_configuration_id` has been set as computed only** and
cannot be set by the user anymore. For now, this change is present on
the Elasticsearch topology only, in the future, as other applications
might also auto scale, the API will change accordingly.
Last, it validates that a set topology element has a size > 0, otherwise
an error will be thrown.

Another somewhat breaking change is the removal of the version computed
property in the Elasticsearch object. This field is completely redundant
and should not be used, instead the global version declares the version
for all of the deployment's applications. Since we're removing a field
from the schema and previous versions of the provider have saved it in
the state, we're adding a `StateUpgradeFunc` and bumping the current
schema version to `1`. The schema upgrade function will only be run when
the saved state version is `0` and will only run once, then the state
schema version will be bumped to 1.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>
Co-authored-by: alaudazzi <46651782+alaudazzi@users.noreply.github.com>