scripts: Add changelog generator (#257)
Adds a changelog generator inspired on the work Hashicorp has done for
some of their providers. The structure is as follows:

Each change (PR) worth having an entry in the changelog should add a new
file `#PR.txt` as part of the Pull request. Examples on the file format
can be found in the `.changelog` folder.

Also, adds a GitHub action that will run every time a PR is merged and
run `make changelog` and commit any changes should there be any new
entries to the changelog file.

Last, I've ran the changelog generation for the upcoming version and
bumped the current provider version to `v0.1.0`.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>