ec_deployment: Fix bug on upgrade with node_roles
Fixes a bug that prevents deployments from upgrading when they're using
the `node_roles` field in the state.

The reason was that since the "version" field was changing, the template
field that was being chosen was `node_type_*` instead of `node_roles`.

Adds an integration test covering this scenario which was  missing, this
being the main reason why we shipped this bug with `0.2.0`.

Closes #328

Signed-off-by: Marc Lopez <marc5.12@outlook.com>