ec_deployment: Store topology.elasticsearch object
Handles the `plan.topology.[].elasticsearch` object as a computed block
that will be flattened (stored) in the state and expanded when present.

This solves the problem where the UI (used to) set these settings at the
topology level and the Terraform Provided set them at the Elasticsearch
level. It still doesn't allow users to set them at the topology level
since that's not the desired behavior, but with after change it will not
unset those fields with every plan change anymore, resulting in less
disruptive behavior.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>