Modify logic to match desired behavior
Updates the logic to match what the resource should do, in short:

Modifies the `create` function to use a hash of the deployment and the
setting name.
Modifies the `delete` function to unset the value of the keystore so the
keystore setting entry is deleted when sent to the `PATCH` api.
Simplifies the `expandModel` logic and allows it to `json.Unmarshal` the
specified string when it's JSON formatted, otherwise defaults to string.
Simplifies the `read` function to lookup the setting name as the key on
the returned secrets map and to only reset the ID since the setting does
not exist in the backing Elasticsearch cluster anymore.
Updates the schema so that when changing the `deployment_id` field, it
triggers a re-creation and correctly deletes the previous entry from the
previous deployment ID and creates a new entry on the new deployment.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>