scripts: Update provider blocks with version
Adds a new script which will be called after each version bump, which
will update all the `terraform.ec.version` references to use the version
which was just bumped. This script will be called every time a new tag
is pushed and thus will maintain all the versions in sync.

This solves an issue where the version would be updated, but the main
`README.md` and the examples would still point to the version which was
just released.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>