UseTopologyStateForUnknown for master tier.
It is important to use UseTopologyStateForUnknown here, as UseStateForUnknown will also set a state for all children of the topology. However the children should stay unknown and their plan modifiers don't run.

This leads to problems when migrating templates, as the master tier should then not use the state and instead get the new values from the new template.