Bugfix: node-types-validation should handle empty stack version.
If the version is unset, the validation crashes with an error. This can happen in certain scenarios:
When using a reference (e.g. `data.ec_stack.latest.version`) it will not be resolved in a first phase when this validation is called. In such cases the validation should be skipped.

The validation still works as it is called again after the reading phase when the field has been resolved.