feat(serverless): Set the credentials.password as sensitive
When creating a new Serverless project, the password of the ES
credentials is marked as non sensitive. This means that it will be shown
in the terraform output when creating an `output` block to print it.

This commit sets the credentials.password of all Serverless resources to
sensitive, so that it is not even allowed to create a terraform `output`
block without marking it as sensitive.

As an example, for the following block:
```terraform
output "password" {
  value = ec_elasticsearch_project.test.credentials.password
}
```

Before this commit, the `terraform apply` outputs the password. After
this commit, we get the following output:

```
│ Error: Output refers to sensitive values
│
│   on main.tf line 28:
│   28: output "password" {
│
│ To reduce the risk of accidentally exporting sensitive data that was intended to be only internal, Terraform requires that any root module output containing sensitive data be explicitly marked as sensitive, to confirm your intent.
│
│ If you do intend to export this data, annotate the output value as sensitive by adding the following argument:
│     sensitive = true
```